feat: work around get_stacktrace deprecation/removal
This patch introduces a macro and inserts it everywhere we catch errors
and then generatre a stacktrace.

So far the only thing that is a little bit ugly is that in two places,
I had to add a header include dependency on couch_db.erl where those
modules didn’t have any ties to couchdb/* before, alas. I’d be willing
to duplicate the macros in those modules, if we don’t want the include
dependency.